+++ apps/web/src/islands/issue-list/CreateIssueModal.tsx
@@ -2,7 +2,7 @@ import type { Dispatch, StateUpdater } from "preact/hooks";
 import { PRIORITY_OPTIONS } from "../../utils/issue-utils";
 import type { TaskStatus } from "../board-utils";
 import MarkdownEditor from "../LazyMarkdownEditor";
-import Select from "../Select";
+import Select from "../ui/Select";
 import type { ProjectMeta } from "./types";
 
 interface CreateIssueModalProps {
